Relegated plot related items to utils
